# PL9/sic-monolith-to-microservices-exercise-pl-08-solution-compose-swarm/microservices-movie-app/reviews-service/swagger-output.json
{
  "swagger": "2.0",
  "info": {
    "title": "My API",
    "description": "Swagger documentation",
    "version": "1.0.0"
  },
  "host": "localhost:3002",
  "basePath": "/",
  "tags": [
    {
      "name": "Reviews",
      "description": "Reviews related endpoints"
    }
  ],
  "schemes": [
    "http"
  ],
  "paths": {
    "/reviews": {
      "get": {
        "tags": [
          "Reviews"
        ],
        "description": "",
        "parameters": [
          {
            "name": "movieId",
            "in": "query",
            "type": "string"
          }
        ],
        "responses": {
          "200": {
            "description": "List of all reviews",
            "schema": {
              "type": "array",
              "items": {
                "$ref": "#/definitions/GetReview"
              },
              "xml": {
                "name": "main"
              }
            }
          }
        }
      },
      "post": {
        "tags": [
          "Reviews"
        ],
        "description": "",
        "parameters": [
          {
            "name": "authorization",
            "in": "header",
            "type": "string"
          },
          {
            "name": "body",
            "in": "body",
            "description": "New review object",
            "required": true,
            "schema": {
              "$ref": "#/definitions/CreateReview"
            }
          }
        ],
        "responses": {
          "201": {
            "description": "Review created successfully",
            "schema": {
              "$ref": "#/definitions/GetReview"
            }
          },
          "401": {
            "description": "Unauthorized"
          },
          "403": {
            "description": "Forbidden"
          }
        }
      }
    },
    "/reviews/{id}": {
      "get": {
        "tags": [
          "Reviews"
        ],
        "description": "",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "type": "string",
            "description": "ID of the review"
          }
        ],
        "responses": {
          "200": {
            "description": "Single review",
            "schema": {
              "$ref": "#/definitions/GetReview"
            }
          },
          "404": {
            "description": "Review not found"
          }
        }
      },
      "put": {
        "tags": [
          "Reviews"
        ],
        "description": "",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "type": "string",
            "description": "ID of the review"
          },
          {
            "name": "body",
            "in": "body",
            "description": "Updated review object",
            "required": true,
            "schema": {
              "$ref": "#/definitions/CreateReview"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Review updated successfully",
            "schema": {
              "$ref": "#/definitions/GetReview"
            }
          },
          "404": {
            "description": "Review not found"
          }
        }
      },
      "delete": {
        "tags": [
          "Reviews"
        ],
        "description": "",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "type": "string",
            "description": "ID of the review"
          },
          {
            "name": "authorization",
            "in": "header",
            "type": "string"
          }
        ],
        "responses": {
          "204": {
            "description": "Review deleted successfully, no content"
          },
          "401": {
            "description": "Unauthorized"
          },
          "403": {
            "description": "Forbidden"
          },
          "404": {
            "description": "Review not found"
          }
        }
      }
    }
  },
  "definitions": {
    "GetReview": {
      "type": "object",
      "properties": {
        "id": {
          "type": "number",
          "example": 1
        },
        "movieId": {
          "type": "string",
          "example": "123"
        },
        "userId": {
          "type": "string",
          "example": "123"
        },
        "text": {
          "type": "string",
          "example": "Review text"
        }
      }
    },
    "CreateReview": {
      "type": "object",
      "properties": {
        "movieId": {
          "type": "string",
          "example": "123"
        },
        "text": {
          "type": "string",
          "example": "Review text"
        }
      }
    }
  }
}